PN-44 #done #comment implement method to award points for feature interaction

+++ naturalist/src/main/java/com/pocket/naturalist/dto/CheckInResponseFeatureDTO.java
@@ -5,6 +5,6 @@ package com.pocket.naturalist.dto;
  * @param featureId
  * @param isNear
  */
-public record CheckInResponseFeatureDTO(String parkSlug, int featureId, boolean isNear) {
+public record CheckInResponseFeatureDTO(String parkSlug, long featureId, boolean isNear) {
     
 }
